// src/app/models/book.interface.ts
export interface BookI {
    idbook?: string;
    title?: string;
    image?: string;
    pdfname?: string;
    nameauthor?: string;
    lastnameauthor?: string;
    editorial?: string;
    categories?: string;
    description?: string;
    edition?: number;
    publicationDate?: Date;
    state?:boolean;
  }
  